refactor(settings): extract SAW + flow-cal into SettingsCalibration domain (#860)
Tier 3 of the Settings domain split. Moves ~25 calibration/SAW-learning
methods, the in-memory caches, and the file-scope constants from the
parent Settings class into a new SettingsCalibration sub-object so the
calibration subsystem can iterate without busting the rebuild blast
across every Settings consumer.

- settings.h: 311 → 200 lines (target met)
- settings.cpp: 1565 → 649 lines (916 lines moved verbatim)
- New settings_calibration.{h,cpp}: 146 + 930 lines
- Storage keys (saw/learningHistory, saw/perProfileHistory,
  saw/perProfileBatch, saw/globalBootstrapLag/<scale>,
  calibration/flowMultiplier, calibration/autoFlowCalibration,
  calibration/perProfileFlow, calibration/flowCalBatch) byte-identical
- resetSawLearning → SettingsBrew hot-water-SAW reset re-wired through a
  sawLearningResetRequested signal + connect() in Settings::Settings()
- FlowCalibrationModel narrowed to SettingsCalibration*
- 1843/1843 tests pass
