+++ tests/CMakeLists.txt
@@ -54,6 +54,7 @@ set(CORE_SOURCES
     ${CMAKE_SOURCE_DIR}/src/core/settings_dye.cpp
     ${CMAKE_SOURCE_DIR}/src/core/settings_network.cpp
     ${CMAKE_SOURCE_DIR}/src/core/settings_app.cpp
+    ${CMAKE_SOURCE_DIR}/src/core/settings_calibration.cpp
 )
 
 set(CONTROLLER_SOURCES
